6.5 payment forms with stripe

+++ db/migrate/20180118095326_change_data_type_for_price.rb
@@ -1,5 +1,5 @@
 class ChangeDataTypeForPrice < ActiveRecord::Migration[5.1]
   def change
-    change_column :products, :price, :float
+    change_column :products, :price, :decimal
   end
 end
